+++ src/Filterable.ts
@@ -1,47 +1,13 @@
 /**
  * `Filterable` represents data structures which can be _partitioned_/_filtered_.
  *
- * Adapted from https://github.com/LiamGoodacre/purescript-filterable/blob/master/src/Data/Filterable.purs
- *
  * @since 2.0.0
  */
-import {
-  Compactable,
-  Compactable1,
-  Compactable2,
-  Compactable2C,
-  Compactable3,
-  CompactableComposition,
-  CompactableComposition11,
-  CompactableComposition12,
-  CompactableComposition12C,
-  CompactableComposition21,
-  CompactableComposition22,
-  getCompactableComposition,
-  Separated,
-  Compactable4,
-  Compactable3C,
-  CompactableComposition23
-} from './Compactable'
+import { Separated } from './Compactable'
 import { Either } from './Either'
 import { Predicate, Refinement } from './function'
-import {
-  Functor,
-  Functor1,
-  Functor2,
-  Functor2C,
-  Functor3,
-  FunctorComposition,
-  FunctorComposition11,
-  FunctorComposition12,
-  FunctorComposition12C,
-  FunctorComposition21,
-  FunctorComposition22,
-  Functor4,
-  Functor3C,
-  FunctorComposition23
-} from './Functor'
-import { HKT, Kind, Kind2, Kind3, URIS, URIS2, URIS3, URIS4, Kind4 } from './HKT'
+import { Functor, Functor1, Functor2, Functor2C } from './Functor'
+import { HKT, Kind, Kind2, Kind3, Kind4, URIS, URIS2, URIS3, URIS4 } from './HKT'
 import { getLeft, getRight, Option } from './Option'
 
 /**
@@ -64,7 +30,8 @@ export interface Partition<F> {
  * @category type classes
  * @since 2.0.0
  */
-export interface Filterable<F> extends Functor<F>, Compactable<F> {
+export interface Filterable<F> {
+  readonly URI: F
   /**
    * Partition a data structure based on an either predicate.
    */
@@ -103,7 +70,8 @@ export interface Partition1<F extends URIS> {
  * @category type classes
  * @since 2.0.0
  */
-export interface Filterable1<F extends URIS> extends Functor1<F>, Compactable1<F> {
+export interface Filterable1<F extends URIS> {
+  readonly URI: F
   readonly partitionMap: <A, B, C>(fa: Kind<F, A>, f: (a: A) => Either<B, C>) => Separated<Kind<F, B>, Kind<F, C>>
   readonly partition: Partition1<F>
   readonly filterMap: <A, B>(fa: Kind<F, A>, f: (a: A) => Option<B>) => Kind<F, B>
@@ -130,7 +98,8 @@ export interface Partition2<F extends URIS2> {
  * @category type classes
  * @since 2.0.0
  */
-export interface Filterable2<F extends URIS2> extends Functor2<F>, Compactable2<F> {
+export interface Filterable2<F extends URIS2> {
+  readonly URI: F
   readonly partitionMap: <E, A, B, C>(
     fa: Kind2<F, E, A>,
     f: (a: A) => Either<B, C>
@@ -160,7 +129,9 @@ export interface Partition2C<F extends URIS2, E> {
  * @category type classes
  * @since 2.0.0
  */
-export interface Filterable2C<F extends URIS2, E> extends Functor2C<F, E>, Compactable2C<F, E> {
+export interface Filterable2C<F extends URIS2, E> {
+  readonly URI: F
+  readonly _E: E
   readonly partitionMap: <A, B, C>(
     fa: Kind2<F, E, A>,
     f: (a: A) => Either<B, C>
@@ -193,7 +164,8 @@ export interface Partition3<F extends URIS3> {
  * @category type classes
  * @since 2.0.0
  */
-export interface Filterable3<F extends URIS3> extends Functor3<F>, Compactable3<F> {
+export interface Filterable3<F extends URIS3> {
+  readonly URI: F
   readonly partitionMap: <R, E, A, B, C>(
     fa: Kind3<F, R, E, A>,
     f: (a: A) => Either<B, C>
@@ -226,7 +198,9 @@ export interface Partition3C<F extends URIS3, E> {
  * @category type classes
  * @since 2.2.0
  */
-export interface Filterable3C<F extends URIS3, E> extends Functor3C<F, E>, Compactable3C<F, E> {
+export interface Filterable3C<F extends URIS3, E> {
+  readonly URI: F
+  readonly _E: E
   readonly partitionMap: <R, A, B, C>(
     fa: Kind3<F, R, E, A>,
     f: (a: A) => Either<B, C>
@@ -259,7 +233,8 @@ export interface Partition4<F extends URIS4> {
  * @category type classes
  * @since 2.0.0
  */
-export interface Filterable4<F extends URIS4> extends Functor4<F>, Compactable4<F> {
+export interface Filterable4<F extends URIS4> {
+  readonly URI: F
   readonly partitionMap: <S, R, E, A, B, C>(
     fa: Kind4<F, S, R, E, A>,
     f: (a: A) => Either<B, C>
@@ -272,7 +247,7 @@ export interface Filterable4<F extends URIS4> extends Functor4<F>, Compactable4<
 /**
  * @since 2.0.0
  */
-export interface FilterableComposition<F, G> extends FunctorComposition<F, G>, CompactableComposition<F, G> {
+export interface FilterableComposition<F, G> {
   readonly partitionMap: <A, B, C>(
     fa: HKT<F, HKT<G, A>>,
     f: (a: A) => Either<B, C>
@@ -288,9 +263,7 @@ export interface FilterableComposition<F, G> extends FunctorComposition<F, G>, C
 /**
  * @since 2.0.0
  */
-export interface FilterableComposition11<F extends URIS, G extends URIS>
-  extends FunctorComposition11<F, G>,
-    CompactableComposition11<F, G> {
+export interface FilterableComposition11<F extends URIS, G extends URIS> {
   readonly partitionMap: <A, B, C>(
     fa: Kind<F, Kind<G, A>>,
     f: (a: A) => Either<B, C>
@@ -306,9 +279,7 @@ export interface FilterableComposition11<F extends URIS, G extends URIS>
 /**
  * @since 2.0.0
  */
-export interface FilterableComposition12<F extends URIS, G extends URIS2>
-  extends FunctorComposition12<F, G>,
-    CompactableComposition12<F, G> {
+export interface FilterableComposition12<F extends URIS, G extends URIS2> {
   readonly partitionMap: <E, A, B, C>(
     fa: Kind<F, Kind2<G, E, A>>,
     f: (a: A) => Either<B, C>
@@ -324,9 +295,7 @@ export interface FilterableComposition12<F extends URIS, G extends URIS2>
 /**
  * @since 2.0.0
  */
-export interface FilterableComposition12C<F extends URIS, G extends URIS2, E>
-  extends FunctorComposition12C<F, G, E>,
-    CompactableComposition12C<F, G, E> {
+export interface FilterableComposition12C<F extends URIS, G extends URIS2, E> {
   readonly partitionMap: <A, B, C>(
     fa: Kind<F, Kind2<G, E, A>>,
     f: (a: A) => Either<B, C>
@@ -342,9 +311,7 @@ export interface FilterableComposition12C<F extends URIS, G extends URIS2, E>
 /**
  * @since 2.0.0
  */
-export interface FilterableComposition21<F extends URIS2, G extends URIS>
-  extends FunctorComposition21<F, G>,
-    CompactableComposition21<F, G> {
+export interface FilterableComposition21<F extends URIS2, G extends URIS> {
   readonly partitionMap: <E, A, B, C>(
     fa: Kind2<F, E, Kind<G, A>>,
     f: (a: A) => Either<B, C>
@@ -360,9 +327,7 @@ export interface FilterableComposition21<F extends URIS2, G extends URIS>
 /**
  * @since 2.0.0
  */
-export interface FilterableComposition2C1<F extends URIS2, G extends URIS, E>
-  extends FunctorComposition21<F, G>,
-    CompactableComposition21<F, G> {
+export interface FilterableComposition2C1<F extends URIS2, G extends URIS, E> {
   readonly partitionMap: <A, B, C>(
     fa: Kind2<F, E, Kind<G, A>>,
     f: (a: A) => Either<B, C>
@@ -378,9 +343,7 @@ export interface FilterableComposition2C1<F extends URIS2, G extends URIS, E>
 /**
  * @since 2.0.0
  */
-export interface FilterableComposition22<F extends URIS2, G extends URIS2>
-  extends FunctorComposition22<F, G>,
-    CompactableComposition22<F, G> {
+export interface FilterableComposition22<F extends URIS2, G extends URIS2> {
   readonly partitionMap: <FE, GE, A, B, C>(
     fa: Kind2<F, FE, Kind2<G, GE, A>>,
     f: (a: A) => Either<B, C>
@@ -402,9 +365,7 @@ export interface FilterableComposition22<F extends URIS2, G extends URIS2>
 /**
  * @since 2.0.0
  */
-export interface FilterableComposition22C<F extends URIS2, G extends URIS2, E>
-  extends FunctorComposition22<F, G>,
-    CompactableComposition22<F, G> {
+export interface FilterableComposition22C<F extends URIS2, G extends URIS2, E> {
   readonly partitionMap: <FE, A, B, C>(
     fa: Kind2<F, FE, Kind2<G, E, A>>,
     f: (a: A) => Either<B, C>
@@ -423,9 +384,7 @@ export interface FilterableComposition22C<F extends URIS2, G extends URIS2, E>
 /**
  * @since 2.2.0
  */
-export interface FilterableComposition23C<F extends URIS2, G extends URIS3, E>
-  extends FunctorComposition23<F, G>,
-    CompactableComposition23<F, G> {
+export interface FilterableComposition23C<F extends URIS2, G extends URIS3, E> {
   readonly partitionMap: <R, FE, A, B, C>(
     fa: Kind2<F, FE, Kind3<G, R, E, A>>,
     f: (a: A) => Either<B, C>
@@ -481,11 +440,7 @@ export function getFilterableComposition<F extends URIS, G extends URIS>(
 ): FilterableComposition11<F, G>
 export function getFilterableComposition<F, G>(F: Functor<F>, G: Filterable<G>): FilterableComposition<F, G>
 export function getFilterableComposition<F, G>(F: Functor<F>, G: Filterable<G>): FilterableComposition<F, G> {
-  const CC = getCompactableComposition(F, G)
   const FC: FilterableComposition<F, G> = {
-    map: CC.map,
-    compact: CC.compact,
-    separate: CC.separate,
     partitionMap: (fga, f) => {
       const left = FC.filterMap(fga, (a) => getLeft(f(a)))
       const right = FC.filterMap(fga, (a) => getRight(f(a)))