+++ src/Traversable.ts
@@ -12,18 +12,6 @@
  *
  * where `A` is an `Applicative` instance
  *
- * `Traversable` instances should also be compatible with the corresponding `Foldable` instances, in the following sense:
- *
- * ```ts
- * import { getApplicative, make } from 'fp-ts/Const'
- *
- * const A = getApplicative(M)
- *
- * foldMap(M)(xs, f) = traverse(A)(xs, a => make(f(a)))
- * ```
- *
- * where `M` is a `Monoid` instance
- *
  * @since 2.0.0
  */
 import {